[System.Net.Http] Clean up HttpMessageHandler setup
to avoid reflection and not to be recursive

Superseded #15051 and #15922
