--- mcs/class/System.Net.Http/wasm_System.Net.Http.dll.exclude.sources
@@ -1,6 +1,0 @@
-HttpClient.DefaultHandler.cs
-HttpClientHandler.cs
-HttpClientHandler.DefaultHandler.cs
-HttpClientHandler.SocketsHandler.cs
-corefx/SocketsHttpHandler.Mono.cs
-../../../external/corefx/src/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/*.cs
