+++ mcs/class/System.Net.Http/wasm_System.Net.Http.dll.sources
@@ -1,3 +1,34 @@
-#include socketshandler.sources
-HttpClientHandler.wasm.cs
+Assembly/AssemblyInfo.cs
+
+corefx/NetEventSource.Http.cs
+
+System.Net.Http/HttpClientHandler.wasm.cs
 System.Net.Http/HttpClient.wasm.cs
+HttpRequestMessage.Mono.cs
+
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/Headers/*.cs
+
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/ByteArrayContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/ByteArrayHelpers.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/ClientCertificateOption.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/DelegatingHandler.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/FormUrlEncodedContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpClient.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpCompletionOption.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpMessageHandler.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpMessageInvoker.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpMethod.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpParseResult.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpRequestException.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpRequestMessage.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpResponseMessage.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpRuleParser.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/HttpUtilities.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/MessageProcessingHandler.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/MultipartContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/MultipartFormDataContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/ReadOnlyMemoryContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/StreamContent.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/StreamToStreamCopy.cs
+../../../external/corefx/src/System.Net.Http/src/System/Net/Http/StringContent.cs
